--- frontend/public/images/trumpet-right.png:Zone.Identifier
@@ -1,4 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=https://www.cleanpng.com/
-HostUrl=https://png2.cleanpng.com/dy/47fd0a90539021d688e5cf103e2f4ed3/L0KzQYm3U8I2N5d4j5H0aYP2gLBuTgBieqVAReJ4cIDogn7qjB5nbaV5gZ98dHBme373iP91d5h3eeJxeT3mfLr3TfFzfF5oh9DvZYT3eX68gfI5PGI7TqYBMEG4RXA9UMg2O2U3T6MAMkKzQoWBUMYzQGc9RuJ3Zx==/kisspng-party-popper-confetti-stock-photography-clip-art-confetti-5ab84166460155.6085342715220248062868.png
