--- frontend/src/components/DiveSpot/DiveSpot.css
@@ -1,35 +1,0 @@
-
-.diveSpotWrapper{
-    margin-top: 20px;
-    background: rgb(0,0,0);
-    background: linear-gradient(0deg, rgba(0,0,0,1) 0%, rgba(40,40,40,1) 59%, rgba(78,78,78,1) 100%); 
-    color: white;
-    border: 4px inset black;
-    width: 75%;
-    height: 85%;
-    display: flex;
-    flex-direction: column;
-    overflow: scroll;
-    overflow-x: hidden;
-}
-
-
-
-.spotList{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-#newSpotForm{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-#newSpotForm > input,
-#newSpotForm > textarea {
-    background: #f1d592;
-}
-
-
