+++ frontend/src/components/NewSpot/NewDiveSpot.js
@@ -4,7 +4,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
 import {createSpot} from '../../store/diveSpot.js'
 
-import './DiveSpot.css';
+import './NewSpot.css';
 
 
 function NewDiveSpotPage(){
