+++ frontend/src/components/SpotDirectory/listing.css
@@ -1,0 +1,5 @@
+.spotList{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
