+++ frontend/src/components/ViewOneSpot/individual.css
