+++ frontend/src/index.css
@@ -13,6 +13,20 @@
     box-sizing: border-box;
 }
 
+.diveSpotWrapper{
+    margin-top: 20px;
+    background: rgb(0,0,0);
+    background: linear-gradient(0deg, rgba(0,0,0,1) 0%, rgba(40,40,40,1) 59%, rgba(78,78,78,1) 100%); 
+    color: white;
+    border: 4px inset black;
+    width: 75%;
+    height: 85%;
+    display: flex;
+    flex-direction: column;
+    overflow: scroll;
+    overflow-x: hidden;
+}
+
 body {
     margin: 0;
     padding: 0;
